requirements.txt
disnake==2.12
Flask==3.0.3
python-dotenv==1.2.2
Requests==2.33